minor 'refactoring'

+++ src/GpsEvent.java
@@ -5,7 +5,7 @@ public class GpsEvent {
   public double latitude;     // The Latitude of the GPS event as a value from -90.0 to +90.0
   public double longitude;    // The Longitude of the GPS event as a value from -180.0 to +180.0
   public double altitude;     // The Altitude of the GPS event in feet
-  public Long timeAdded;
+  public Double timeAdded;
 
   /** Creates a GpsEvent. */
   public GpsEvent(String name, double latitude, double longitude, double altitude){
@@ -20,7 +20,7 @@ public class GpsEvent {
    * 
    * @param time  The time the event was added to the cell list in DistanceTravelledComponent.
    */
-  public void setTime(Long time){
+  public void setTime(Double time){
     this.timeAdded = time;
   }
